ClassRoom separado do ClassRoomBoard
