+++ Assets/ClassRoom.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: de4e28f20d17910419aeb6e7a8cbe116
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
